refactor: split gateway runtime manager
