+++ crates/omniinfer-cli/src/gateway/tests.rs
@@ -1,3 +1,4 @@
+use super::runtime_manager::{LoadedRuntimeSummary, pick_runtime_port};
 use super::*;
 
 static TEST_ENV_LOCK: tokio::sync::Mutex<()> = tokio::sync::Mutex::const_new(());
